Add support for logs level.
The debug, info, warn, error and fatal functions now directly
pass the passed-in string to the logger.
The console logging backend supports the level of a log.
The verbose environment variable is replaced by a verbosity variable
that defines the minimum level of logs to display.
Commands that have to output something to the console now use
console.log rather than puts.