Add travis status in README.
This push will act as a test of travis use with nodejs.
All the tests passed on my computer, if it failed now, even though
I disabled functional tests that was systematically failing on travis,
it will clearly state that travis can't be relied on when running tests
for nodejs code. It's worth noting that the failing tests was ran
hundred of times without failures before using travis.